+++ packages/kysely/test/users/index.spec.ts
@@ -8,7 +8,7 @@ describe("users", () => {
 
     await data.tenants.create({
       id: "tenantId",
-      name: "Test Tenant",
+      friendly_name: "Test Tenant",
       audience: "https://example.com",
       sender_email: "login@example.com",
       sender_name: "SenderName",